Add Multiline Regex Data Capture log filter functional test
